requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
sentence-transformers
chromadb
coqui-tts
sounddevice
numpy
groq
python-dotenv
webrtcvad
pypdf
edge-tts
gtts
pygame
pyttsx3
TTS
soundfile
scipy